+++ src/Mod/Draft/draftguitools/gui_selectplane.py
@@ -494,44 +494,4 @@ class Draft_SelectPlane:
         FreeCADGui.doCommandGui("FreeCADGui.Snapper.setGrid()")
 
 
-class Draft_WorkingPlaneProxy:
-    """The Draft_WorkingPlaneProxy command definition."""
-
-    def GetResources(self):
-        """Set icon, menu and tooltip."""
-        _menu = "Create working plane proxy"
-        _tip = ("Creates a proxy object from the current working plane.\n"
-                "Once the object is created double click it in the tree view "
-                "to restore the camera position and objects' visibilities.\n"
-                "Then you can use it to save a different camera position "
-                "and objects' states any time you need.")
-        d = {'Pixmap': 'Draft_PlaneProxy',
-             'MenuText': QT_TRANSLATE_NOOP("Draft_SetWorkingPlaneProxy",
-                                           _menu),
-             'ToolTip': QT_TRANSLATE_NOOP("Draft_SetWorkingPlaneProxy",
-                                          _tip)}
-        return d
-
-    def IsActive(self):
-        """Return True when this command should be available."""
-        if FreeCADGui.ActiveDocument:
-            return True
-        else:
-            return False
-
-    def Activated(self):
-        """Execute when the command is called."""
-        if hasattr(FreeCAD, "DraftWorkingPlane"):
-            FreeCAD.ActiveDocument.openTransaction("Create WP proxy")
-            FreeCADGui.addModule("Draft")
-            _cmd = "Draft.makeWorkingPlaneProxy("
-            _cmd += "FreeCAD.DraftWorkingPlane.getPlacement()"
-            _cmd += ")"
-            FreeCADGui.doCommand(_cmd)
-            FreeCAD.ActiveDocument.commitTransaction()
-            FreeCAD.ActiveDocument.recompute()
-
-
 FreeCADGui.addCommand('Draft_SelectPlane', Draft_SelectPlane())
-FreeCADGui.addCommand('Draft_WorkingPlaneProxy',
-                      Draft_WorkingPlaneProxy())
